refactor: extract shared parse markdown
- Added `@microflash/remark-callout-directives` to `package.json` and `pnpm-lock.yaml`.
- Updated `MarkdownWeb` component to utilize a new markdown parsing utility.
- Removed unused imports related to `remark-gh-alerts` and refactored markdown parsing logic.
- Enhanced `parse-markdown` function to integrate with the new utility and support custom components.

This update improves markdown rendering capabilities and streamlines the codebase.
