+++ packages/components/package.json
@@ -28,6 +28,7 @@
     "@follow/types": "workspace:*",
     "@follow/utils": "workspace:*",
     "@headlessui/react": "2.2.0",
+    "@microflash/remark-callout-directives": "4.3.2",
     "@radix-ui/react-avatar": "1.1.1",
     "@radix-ui/react-checkbox": "1.1.2",
     "@radix-ui/react-context-menu": "2.2.2",
@@ -65,6 +66,10 @@
     "rehype-parse": "9.0.1",
     "rehype-sanitize": "6.0.0",
     "rehype-stringify": "10.0.1",
+    "remark-directive": "3.0.0",
+    "remark-gfm": "4.0.0",
+    "remark-gh-alerts": "0.0.3",
+    "remark-rehype": "11.1.1",
     "sonner": "1.7.1",
     "unified": "11.0.5",
     "unist-util-visit": "5.0.0",
